Move DebugSpawn(Unique)Item to lua dev moduls and make necessary functions public
